+++ src/test_data/test_client_impl_err.rs
@@ -62,7 +62,7 @@ impl KrakenClient for TestClientErr {
     fn new_with_url(
         _secrets_provider: Box<Arc<Mutex<dyn SecretsProvider>>>,
         _nonce_provider: Box<Arc<Mutex<dyn NonceProvider>>>,
-        _url: String,
+        _url: impl ToString,
     ) -> Self {
         Self {}
     }
@@ -75,7 +75,7 @@ impl KrakenClient for TestClientErr {
         Self {}
     }
 
-    async fn set_user_agent(&mut self, _user_agent: String) {}
+    async fn set_user_agent(&mut self, _user_agent: impl ToString) {}
 
     async fn get_server_time(&mut self) -> Result<ResultErrorResponse<SystemTime>, ClientError> {
         Err(ClientError::Parse("StubbedForTesting"))