+++ src/wss/kraken_wss_client.rs
@@ -11,7 +11,7 @@ use tokio::net::TcpStream;
 use tokio_stream::Stream;
 use tokio_tungstenite::tungstenite::Message as TungsteniteMessage;
 use tokio_tungstenite::{connect_async, MaybeTlsStream, WebSocketStream};
-use tracing::{debug, trace, warn};
+use tracing::trace;
 use url::Url;
 
 pub const WS_KRAKEN: &str = "wss://ws.kraken.com/v2";
@@ -37,37 +37,19 @@ impl Default for KrakenWSSClient {
 impl KrakenWSSClient {
     /// Create a client using the default Kraken URLs.
     pub fn new() -> KrakenWSSClient {
-        if cfg!(feature = "debug-inbound") {
-            warn!("Feature `debug-inbound` is deprecated - please use `new_with_tracing` method on `KrakenWSSClient`")
-        }
-
-        if cfg!(feature = "debug-outbound") {
-            warn!("Feature `debug-outbound` is deprecated - please use `new_with_tracing` method on `KrakenWSSClient`")
-        }
-
-        KrakenWSSClient {
-            base_url: WS_KRAKEN.to_string(),
-            auth_url: WS_KRAKEN_AUTH.to_string(),
-            trace_inbound: false,
-            trace_outbound: false,
-        }
+        KrakenWSSClient::new_with_tracing(WS_KRAKEN, WS_KRAKEN_AUTH, false, false)
     }
 
     /// Create a client with custom URLs.
     ///
     /// This is most useful for use with a proxy, or for testing.
-    pub fn new_with_urls(base_url: String, auth_url: String) -> KrakenWSSClient {
-        KrakenWSSClient {
-            base_url,
-            auth_url,
-            trace_inbound: false,
-            trace_outbound: false,
-        }
+    pub fn new_with_urls(base_url: impl ToString, auth_url: impl ToString) -> KrakenWSSClient {
+        KrakenWSSClient::new_with_tracing(base_url, auth_url, false, false)
     }
 
     pub fn new_with_tracing(
-        base_url: &str,
-        auth_url: &str,
+        base_url: impl ToString,
+        auth_url: impl ToString,
         trace_inbound: bool,
         trace_outbound: bool,
     ) -> KrakenWSSClient {
@@ -164,10 +146,6 @@ where
     {
         let message_json = serde_json::to_string(message)?;
 
-        if cfg!(feature = "debug-outbound") {
-            debug!("Sending: {}", message_json);
-        }
-
         if self.trace_outbound {
             trace!("Sending: {}", message_json);
         }
@@ -188,9 +166,6 @@ where
     /// returns Poll:Ready with a message if available, otherwise Poll:Pending
     fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
         if let Poll::Ready(Some(message)) = Pin::new(&mut self.stream).poll_next(cx)? {
-            if cfg!(feature = "debug-inbound") {
-                trace!("Received: {}", message.to_string());
-            }
             if self.trace_inbound {
                 trace!("Received: {}", message.to_string());
             }
@@ -304,9 +279,7 @@ mod tests {
     async fn test_tracing_flags_enabled() {
         let mock_server = WsMockServer::start().await;
         let uri = mock_server.uri().await;
-        let mut client = KrakenWSSClient::new_with_urls(uri.clone(), uri);
-        client.trace_inbound = true;
-        client.trace_outbound = true;
+        let mut client = KrakenWSSClient::new_with_tracing(&uri, &uri, true, true);
 
         WsMock::new()
             .matcher(Any::new())